move e2e tests to separate directory to avoid nightwatch interpreting page objects as tests
